changed longs to ints because it's probably better

+++ Scale.cpp
@@ -12,27 +12,27 @@ void Scale::init(){
     hx711_scale.tare();
 }
 
-long Scale::read(){
+int Scale::read(){
     if(hx711_scale.is_ready()){
-        long reading = hx711_scale.get_units();
+        int reading = hx711_scale.get_units();
         return reading;
     }
     else return 0;
 }
 
-long Scale::read(byte num_readings){
+int Scale::read(byte num_readings){
     if(hx711_scale.is_ready()){
-        long reading = hx711_scale.get_units(num_readings);
+        int reading = hx711_scale.get_units(num_readings);
         return reading;
     }
     else return 0;
 }
 
 byte Scale::getWeight(){
-    long reading = read(5);
-    long lightDistance = abs(reading - LIGHT_AVG);
-    long mediumDistance = abs(reading - MEDIUM_AVG);
-    long heavyDistance = abs(reading - HEAVY_AVG);
+    int reading = read(5);
+    int lightDistance = abs(reading - LIGHT_AVG);
+    int mediumDistance = abs(reading - MEDIUM_AVG);
+    int heavyDistance = abs(reading - HEAVY_AVG);
     if(lightDistance < mediumDistance && lightDistance < heavyDistance){
         return LIGHT;
     }
